requirements.txt
astrapy==1.5.2
Flask==3.1.0
flask_cors==5.0.0
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
